# server/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "BISense API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/compare": {
   "post": {
    "tags": [
     "misc"
    ],
    "summary": "Compare",
    "operationId": "compare_api_compare_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompareRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CompareResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/voice/tts": {
   "post": {
    "tags": [
     "misc"
    ],
    "summary": "Voice Tts",
    "operationId": "voice_tts_api_voice_tts_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TTSRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/debug/trace/{request_id}": {
   "get": {
    "tags": [
     "misc"
    ],
    "summary": "Debug Trace",
    "operationId": "debug_trace_api_debug_trace__request_id__get",
    "parameters": [
     {
      "name": "request_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Request Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AskTrace"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AskTrace": {
    "properties": {
     "request_id": {
      "type": "string",
      "title": "Request Id"
     },
     "query": {
      "$ref": "#/components/schemas/QueryInfo"
     },
     "candidates": {
      "items": {
       "$ref": "#/components/schemas/Citation"
      },
      "type": "array",
      "title": "Candidates"
     },
     "cited_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Cited Ids"
     },
     "drops": {
      "items": {
       "$ref": "#/components/schemas/Drop"
      },
      "type": "array",
      "title": "Drops"
     },
     "mode": {
      "type": "string",
      "enum": [
       "live",
       "cached",
       "extractive",
       "none"
      ],
      "title": "Mode"
     },
     "provider": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Provider"
     },
     "timings": {
      "$ref": "#/components/schemas/Timings"
     },
     "top_rerank": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Top Rerank"
     },
     "gate_threshold": {
      "type": "number",
      "title": "Gate Threshold"
     },
     "tokens": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Tokens"
     }
    },
    "type": "object",
    "required": [
     "request_id",
     "query",
     "candidates",
     "cited_ids",
     "drops",
     "mode",
     "provider",
     "timings",
     "top_rerank",
     "gate_threshold",
     "tokens"
    ],
    "title": "AskTrace",
    "description": "Everything the Retrieval details drawer shows (also served by /api/debug/trace when DEBUG=true)."
   },
...
   "Citation": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "n": {
      "type": "integer",
      "title": "N"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "standard_number": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Standard Number"
     },
     "standard_title": {
      "type": "string",
      "title": "Standard Title"
     },
     "standard_kind": {
      "type": "string",
      "enum": [
       "standard",
       "guidance",
       "order",
       "catalogue"
      ],
      "title": "Standard Kind"
     },
     "clause_number": {
      "type": "string",
      "title": "Clause Number"
     },
     "clause_heading": {
      "type": "string",
      "title": "Clause Heading"
     },
     "clause_kind": {
      "type": "string",
      "title": "Clause Kind"
     },
     "clause_path": {
      "type": "string",
      "title": "Clause Path"
     },
     "page_start": {
      "type": "integer",
      "title": "Page Start"
     },
     "page_end": {
      "type": "integer",
      "title": "Page End"
     },
     "snippet": {
      "type": "string",
      "title": "Snippet"
     },
     "scores": {
      "$ref": "#/components/schemas/Scores"
     },
     "synthetic": {
      "type": "boolean",
      "title": "Synthetic"
     },
     "tier": {
      "type": "string",
      "title": "Tier"
     },
     "doc_type": {
      "type": "string",
      "title": "Doc Type"
     },
     "url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Url"
     },
     "has_page_image": {
      "type": "boolean",
      "title": "Has Page Image"
     }
    },
    "type": "object",
    "required": [
     "id",
     "n",
     "slug",
     "standard_number",
     "standard_title",
     "standard_kind",
     "clause_number",
     "clause_heading",
     "clause_kind",
     "clause_path",
     "page_start",
     "page_end",
     "snippet",
     "scores",
     "synthetic",
     "tier",
     "doc_type",
     "url",
     "has_page_image"
    ],
    "title": "Citation"
   },
...
   "CompareCell": {
    "properties": {
     "text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Text"
     },
     "citations": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Citations"
     },
     "found": {
      "type": "boolean",
      "title": "Found"
     }
    },
    "type": "object",
    "required": [
     "text",
     "citations",
     "found"
    ],
    "title": "CompareCell"
   },
   "CompareRequest": {
    "properties": {
     "a": {
      "type": "string",
      "pattern": "^[a-z0-9][a-z0-9-]{0,120}$",
      "title": "A"
     },
     "b": {
      "type": "string",
      "pattern": "^[a-z0-9][a-z0-9-]{0,120}$",
      "title": "B"
     },
     "lang": {
      "type": "string",
      "enum": [
       "en",
       "hi",
       "kn"
      ],
      "title": "Lang",
      "default": "en"
     }
    },
    "type": "object",
    "required": [
     "a",
     "b"
    ],
    "title": "CompareRequest"
   },
   "CompareResponse": {
    "properties": {
     "a": {
      "$ref": "#/components/schemas/StandardSummary"
     },
     "b": {
      "$ref": "#/components/schemas/StandardSummary"
     },
     "rows": {
      "items": {
       "$ref": "#/components/schemas/CompareRow"
      },
      "type": "array",
      "title": "Rows"
     },
     "numeric": {
      "items": {
       "$ref": "#/components/schemas/NumericRow"
      },
      "type": "array",
      "title": "Numeric"
     },
     "key_differences": {
      "items": {
       "$ref": "#/components/schemas/Point"
      },
      "type": "array",
      "title": "Key Differences"
     },
     "citations": {
      "items": {
       "$ref": "#/components/schemas/Citation"
      },
      "type": "array",
      "title": "Citations"
     },
     "mode": {
      "type": "string",
      "enum": [
       "live",
       "cached",
       "extractive",
       "none"
      ],
      "title": "Mode"
     },
     "notice": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notice"
     },
     "dropped_count": {
      "type": "integer",
      "title": "Dropped Count",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "a",
     "b",
     "rows",
     "numeric",
     "key_differences",
     "citations",
     "mode",
     "notice",
     "dropped_count"
    ],
    "title": "CompareResponse"
   },
   "CompareRow": {
    "properties": {
     "aspect": {
      "type": "string",
      "title": "Aspect"
     },
     "a": {
      "$ref": "#/components/schemas/CompareCell"
     },
     "b": {
      "$ref": "#/components/schemas/CompareCell"
     }
    },
    "type": "object",
    "required": [
     "aspect",
     "a",
     "b"
    ],
    "title": "CompareRow"
   },
...
   "Drop": {
    "properties": {
     "field": {
      "type": "string",
      "title": "Field"
     },
     "text": {
      "type": "string",
      "title": "Text"
     },
     "reason": {
      "type": "string",
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "field",
     "text",
     "reason"
    ],
    "title": "Drop"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "NumericRow": {
    "properties": {
     "parameter": {
      "type": "string",
      "title": "Parameter"
     },
     "unit": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Unit"
     },
     "a_value": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "A Value"
     },
     "b_value": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "B Value"
     },
     "a_citation": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "A Citation"
     },
     "b_citation": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "B Citation"
     }
    },
    "type": "object",
    "required": [
     "parameter",
     "unit",
     "a_value",
     "b_value",
     "a_citation",
     "b_citation"
    ],
    "title": "NumericRow"
   },
...
   "Point": {
    "properties": {
     "kind": {
      "type": "string",
      "enum": [
       "source_fact",
       "interpretation"
      ],
      "title": "Kind"
     },
     "text": {
      "type": "string",
      "title": "Text"
     },
     "citations": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Citations"
     },
     "quote": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quote"
     }
    },
    "type": "object",
    "required": [
     "kind",
     "text",
     "citations",
     "quote"
    ],
    "title": "Point"
   },
...
   "QueryInfo": {
    "properties": {
     "interpreted_query": {
      "type": "string",
      "title": "Interpreted Query"
     },
     "intent": {
      "type": "string",
      "title": "Intent"
     },
     "lang": {
      "type": "string",
      "enum": [
       "en",
       "hi",
       "kn"
      ],
      "title": "Lang"
     },
     "resolved_scope": {
      "items": {
       "$ref": "#/components/schemas/ScopeItemOut"
      },
      "type": "array",
      "title": "Resolved Scope"
     },
     "scope_source": {
      "type": "string",
      "title": "Scope Source"
     },
     "rewritten": {
      "type": "boolean",
      "title": "Rewritten"
     },
     "notes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "interpreted_query",
     "intent",
     "lang",
     "resolved_scope",
     "scope_source",
     "rewritten",
     "notes"
    ],
    "title": "QueryInfo"
   },
...
   "ScopeItemOut": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "number": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Number"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "kind": {
      "type": "string",
      "enum": [
       "standard",
       "guidance",
       "order",
       "catalogue"
      ],
      "title": "Kind"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "number",
     "title",
     "kind"
    ],
    "title": "ScopeItemOut"
   },
   "Scores": {
    "properties": {
     "lexical_rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lexical Rank"
     },
     "vector": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vector"
     },
     "fused": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fused"
     },
     "rerank": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rerank"
     },
     "boosts": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Boosts"
     }
    },
    "type": "object",
    "required": [
     "lexical_rank",
     "vector",
     "fused",
     "rerank",
     "boosts"
    ],
    "title": "Scores"
   },
...
   "StandardSummary": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "kind": {
      "type": "string",
      "enum": [
       "standard",
       "guidance",
       "order",
       "catalogue"
      ],
      "title": "Kind"
     },
     "number": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Number"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "year": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Year"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "compulsory": {
      "type": "string",
      "title": "Compulsory"
     },
     "catalogue_only": {
      "type": "boolean",
      "title": "Catalogue Only"
     },
     "synthetic": {
      "type": "boolean",
      "title": "Synthetic"
     },
     "needs_review": {
      "type": "boolean",
      "title": "Needs Review"
     },
     "tier": {
      "type": "string",
      "title": "Tier"
     },
     "clause_count": {
      "type": "integer",
      "title": "Clause Count",
      "default": 0
     },
     "requirement_count": {
      "type": "integer",
      "title": "Requirement Count",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "slug",
     "kind",
     "number",
     "title",
     "year",
     "category",
     "status",
     "compulsory",
     "catalogue_only",
     "synthetic",
     "needs_review",
     "tier",
     "clause_count",
     "requirement_count"
    ],
    "title": "StandardSummary"
   },
...
   "TTSRequest": {
    "properties": {
     "text": {
      "type": "string",
      "maxLength": 3000,
      "minLength": 1,
      "title": "Text"
     },
     "lang": {
      "type": "string",
      "enum": [
       "en",
       "hi",
       "kn"
      ],
      "title": "Lang",
      "default": "en"
     }
    },
    "type": "object",
    "required": [
     "text"
    ],
    "title": "TTSRequest"
   },
   "Timings": {
    "properties": {
     "stages": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Stages"
     },
     "total_ms": {
      "type": "number",
      "title": "Total Ms"
     }
    },
    "type": "object",
    "required": [
     "stages",
     "total_ms"
    ],
    "title": "Timings"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}